added functionality
functionality of slider, I managed to use my basic data structure knowledge to use it for slider productivity, and that is - circular indexing, no matter how many times u click the slider to go forward - if it passes it's row barrier it will get back to it's starting place by using module parameter. I'll talk about it more when I'll present it.